Redme
# COVID-19 Tracker

### [Live Site](https://covid19statswebsite.netlify.com/)

![COVID-19 Tracker](https://i.ibb.co/X87BqVY/Screenshot-2020-04-13-at-10-14-58.png)

## Stay up to date with new projects
New major projects coming soon, subscribe to the mailing list to stay up to date https://javascriptmastery.eo.page/mailing-list.

## Introduction
This is a code repository for the corresponding video tutorial. 

In this video, we will create a full COVID-19 Tracker. We're going to use React, Charts.JS and Material UI.

By the end of this video, you will have a strong understanding of React's workflow and the use of hooks.

API used: https://covid19.mathdro.id/api

Setup:
- run ```npm i && npm start```